lib/requirements_ext.txt
pydot==1.4.1
jupyter==1.0.0
torch==1.10.1
torchvision==0.11.2
pycocotools-fix==2.0.0.1
numba==0.48.0
librosa==0.8.0
networkx==2.5
python-magic==0.4.18
beautifulsoup4==4.9.2
cssutils==1.0.2
premailer==3.7.0
fasttext==0.9.2
faster-fifo==1.0.7
tensorflow-hub==0.12.0
tensorflow-text==2.5.0
regex==2020.11.13
langdetect==1.0.8
warcio==1.7.4
tldextract==3.1.0
tensorboard-plugin-profile==2.4.0
spafe==0.1.2
ftfy==6.0.3
transformers==4.15.0
sentence-transformers==2.1.0
gensim==4.1.2
nltk==3.6.3
keras-tuner==1.1.2
brotli==1.0.9
